+++ dorm/cli.py
@@ -405,7 +405,7 @@ def cmd_shell(args):
     banner = "djanorm interactive shell\nModels: " + ", ".join(sorted(models.keys()))
 
     try:
-        import IPython  # type: ignore
+        import IPython
 
         IPython.embed(user_ns=local_vars, banner1=banner, using="asyncio")
         return